fix imports / relocate PytestStep to tests/python_connectors.py
